+++ app/src/main/java/com/example/passbook/daos/IDAO.java
@@ -6,7 +6,7 @@ public interface IDAO<T> {
     public List<T> getItems();
     public long insertItem(T item);
     public void updateOrInsertItem(T item);
-    public T getItem(int id);
+    public T getItem(long id);
     public int deleteItem(T item);
     public void deleteAll();
 }
